separated each class into its own folder
